+++ clients/src/misc/constants.js
@@ -1,0 +1,2 @@
+export const serverBase = process.env.REACT_APP_SS_SERVER_URL
+export const clientBase = process.env.REACT_APP_SS_CLIENT_URL